Remove duplication in reprs
Performed Extract Superclass refactoring on _SequenceReprBuilder
and _TabularMSAReprBuilder. Extracted superclass
_MetadataReprBuilder and moved duplicate functionality there.
Also performed multiple Extract Method refactorings for
subsections of the build method, with some of them becoming
Template Methods.
